+++ eslint.config.js
@@ -9,7 +9,7 @@ import tseslint from 'typescript-eslint'
 import { defineConfig, globalIgnores } from 'eslint/config'
 
 export default defineConfig([
-  globalIgnores(['dist', 'storybook-static']),
+  globalIgnores(['dist', 'storybook-static', 'tests/bundle/**']),
   {
     files: ['**/*.{ts,tsx}'],
     extends: [